// packages/api/src/services/farcaster/types.ts
export type UserCastsResponse = {
	casts: Array<any>;
	next?: { cursor: string | null };
};